Merge master with staging-api

+++ app/services/sns_topic_manager.rb
@@ -17,26 +17,31 @@ class SnsTopicManager
   def generate
     response = sns_client.create_topic(name: "#{environment_prefix}follow_#{participant.slug}")
     if response.successful?
-      print '.'
+      print '.' unless Rails.env.test?
       Rails.logger.info "Generated SNS topic for #{participant.slug}"
       response.topic_arn.include?('arn:aws:sns') ? response.topic_arn : "#{response.topic_arn}:#{SecureRandom.uuid}"
     else
-      print 'X'
-      Rails.logger.info "Unable to generate SNS topic for #{participant.slug}"
+      print 'X' unless Rails.env.test?
+      warn "Unable to generate SNS topic for #{participant.slug}"
       nil
     end
   end
 
   def delete
-    response = sns_client.delete_topic(topic_arn: topic_arn)
-    if response.successful?
-      print '.'
-      Rails.logger.info "Deleted SNS topic #{topic_arn}"
-      topic_arn
+    if topic_arn && topic_arn.include?('arn:aws:sns')
+      response = sns_client.delete_topic(topic_arn: topic_arn)
+      if response.successful?
+        print '.' unless Rails.env.test?
+        Rails.logger.info "Deleted SNS topic #{topic_arn}"
+        topic_arn
+      else
+        print 'X' unless Rails.env.test?
+        warn "Unable to delete SNS topic #{topic_arn}"
+        nil
+      end
     else
-      print 'X'
-      Rails.logger.info "Unable to delete SNS topic #{topic_arn}"
-      nil
+      print '-' unless Rails.env.test?
+      warn "#{participant.slug} has no topic_resource_key or topic_arn does not exist" unless Rails.env.test?
     end
   end
 
@@ -48,7 +53,6 @@ class SnsTopicManager
     participant.topic_resource_key
   end
 
-# TODO Create separate AWS accounts with separate credentials for dev, staging, and production environments
   def environment_prefix
     @environment_prefix ||= Rails.env.production? ? '' : "#{Rails.env.first}-"
   end
